+++ deriver/error.mli
@@ -1,0 +1,9 @@
+open Ppxlib
+
+val unsupported_mutually_rec_type_decl : loc:Location.t -> extension
+val unsupported_type_param : loc:Location.t -> extension
+val unsupported_longident : loc:Location.t -> extension
+val unsupported_type : loc:Location.t -> extension
+val unsupported_gadt : loc:Location.t -> extension
+val unsupported_rinherit : loc:Location.t -> extension
+val unsupported_conjunctive_tag_arg : loc:Location.t -> extension