+++ CHANGES.md
@@ -3,6 +3,8 @@
 ### Added
 
 - Initial implementation of `ppx_deriving_dyn`
+- Allow overriding the default `Dyn.t` converter for a given type or record
+  field via the `[@to_dyn ...]` attribute
 
 ### Changed